reparatory-justice-backend/requirements.txt
Django>=4.2,<5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-simple-history==3.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
resend==2.1.0

